+++ AGrocery/Assets/Andrew/_Scripts/NonVR/Product3ColliderScript.cs
@@ -50,7 +50,7 @@ public class Product3ColliderScript : MonoBehaviour
         { 
                 if (holdingProduct && Input.GetKeyDown(KeyCode.R))
                 {
-                PlayerMoneyHandler.TotalCost += 2.00m;
+                PlayerMoneyHandler.TotalCost += (float)2.00m;
 
                 totalMoneyText = totalMoneyTextBox.GetComponent<Text>();
                 totalMoneyText.text = "Total Cost: " + PlayerMoneyHandler.TotalCost;
